// apps/web/src/generated/models/rocPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * CardioPredict Heart Disease Prediction API
 * Accepts patient clinical data and predicts heart disease risk.
 * OpenAPI spec version: 2.0.0
 */

export interface RocPoint {
  /** False Positive Rate */
  fpr: number;
  /** True Positive Rate */
  tpr: number;
}